refactor: add useFilmModal and useFilmData hooks for better separation
- Add useFilmModal hook for modal state management
- Rename and refactor useFilmData hook for TMDB API integration
- Apply hooks to both film pages for consistent modal handling
- Remove unused imports from FilmModal component
- Update task progress to reflect completed Phase 1 and Phase 3
