+++ src/hooks/useFilmModal.ts
@@ -1,23 +1,28 @@
-import { useCallback, useState, useEffect } from "react";
-import { TMDB_API_KEY } from "@/config";
-import { TMDB_API_URL } from "@/config/constants";
+import { useDisclosure } from "@chakra-ui/react";
+import { useState } from "react";
 
-export const useFilmModal = (filmId?: number) => {
-    const [filmData, setFilmData] = useState<any>();
+interface UseFilmModalReturn {
+    filmId: number | undefined;
+    isOpen: boolean;
+    onClose: () => void;
+    openModal: (_filmId: number) => void;
+}
 
-    const fetchFilmData = useCallback(async (id: number) => {
-        const response = await fetch(`${TMDB_API_URL}/movie/${id}?language=en-US&api_key=${TMDB_API_KEY}`, { method: "GET" });
-        const data = await response.json();
-        setFilmData(data);
-    }, []);
+export const useFilmModal = (): UseFilmModalReturn => {
+    const [filmId, setFilmId] = useState<number | undefined>();
+    const { isOpen, onOpen, onClose } = useDisclosure();
 
-    useEffect(() => {
-        if (filmId) {
-            fetchFilmData(filmId);
-        }
-    }, [filmId, fetchFilmData]);
+    const openModal = (id: number) => {
+        setFilmId(id);
+        setTimeout(() => {
+            onOpen();
+        }, 50);
+    };
 
     return {
-        filmData,
+        filmId,
+        isOpen,
+        onClose,
+        openModal,
     };
 };
